vs2015 x64 debug/release. changed runtime msvcrt to dll. required for multithreading
git-svn-id: svn://svn.code.sf.net/p/irrlicht/code/branches/mipmap@5713 dfc29bdd-3216-0410-991c-e03cc46cb475

+++ source/Irrlicht/CD3D9HLSLMaterialRenderer.cpp
@@ -193,7 +193,7 @@ bool CD3D9HLSLMaterialRenderer::createHLSLPixelShader(const char* pixelShaderPro
 	// compile without debug info
 	HRESULT h = stubD3DXCompileShader(
 		pixelShaderProgram,
-		strlen(pixelShaderProgram),
+		(UINT)strlen(pixelShaderProgram),
 		0, // macros
 		0, // no includes
 		shaderEntryPointName,
